+++ sock.go
@@ -73,6 +73,24 @@ func GetSockFP() string {
 	return RunPath("webexec.sock")
 }
 
+func (s *sockServer) handlePaste(w http.ResponseWriter, r *http.Request) {
+	// Implementation of handlePaste will be added here
+	Logger.Info("Handling paste request")
+	peer := peers.GetRecentPeer()
+	if peer == nil {
+		http.Error(w, "No active peers", http.StatusServiceUnavailable)
+		return
+	}
+	err, reply := peer.SendControlMessageAndWait("get_clipboard", nil)
+	if err != nil {
+		Logger.Errorf("Failed to send the paste message: %s", err)
+		http.Error(w, "Failed to send the paste message", http.StatusInternalServerError)
+		return
+	}
+	w.Write([]byte(reply))
+	return
+}
+
 func StartSocketServer(lc fx.Lifecycle, s *sockServer) (*http.Server, error) {
 	fp := GetSockFP()
 	_, err := os.Stat(fp)
@@ -98,6 +116,7 @@ func StartSocketServer(lc fx.Lifecycle, s *sockServer) (*http.Server, error) {
 	m.Handle("/status", http.HandlerFunc(s.handleStatus))
 	m.Handle("/layout", http.HandlerFunc(s.handleLayout))
 	m.Handle("/offer/", http.HandlerFunc(s.handleOffer))
+	m.Handle("/paste", http.HandlerFunc(s.handlePaste))
 	server := http.Server{Handler: &m}
 	lc.Append(fx.Hook{
 		OnStart: func(context.Context) error {